Trig collisions - Maybe an option
Two particles (particle 1 is blue and particle 2 is red I'm pretty sure), "colliding" using sin and cos. Basically if the red ball is on the left the x gives -1 and if its on the right it gives +1, same for sin and the y axis. Currently draws a yellow dot at the "point of collision".